requirements.txt
dash>=2.0.0
dash-mantine-components>=1.0.0
pytest>=7.0
psutil>=5.0
PyYAML>=6.0
